1.更换双端app图标；2.添加iOS-plist相关权限；3. 添加左侧抽屉，增加UI展示；4.语言国际化本地化(增加多语言配置文件，集成 flutter_localizations:)；5.暗黑模式；6.主题颜色设置(增加主题颜色选择页面)；7.集成barcode_scan2,首页添加扫码功能；8.集成image_picker,个人主页增加拍照，相册选择功能，头像可以存储在本地；9.集成webview_flutter, 实现关于商城中可以打开外部传入的在线链接网页，并展示；10.其他细节调整；

+++ lib/app/modules/setting/views/setting_view.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:get/get.dart';
@@ -51,7 +52,9 @@ class SettingView extends GetView<SettingController> {
           commonListTile("小米商城隐私政策-简要版", onTap: () {}),
           Container(height: DoScreenAdapter.h(10), color: DoColors.gray238),
           commonListTile("关于商城", onTap: () {
-            // print("关于商城");
+            Get.toNamed("/web", arguments: {
+              "url": "https://book.flutterchina.club/chapter12/flutter_web.html"
+            });
           }),
           commonListTile("网络诊断", onTap: () {}),
           commonListTile("个人信息收集与使用清单", onTap: () {}),
@@ -136,15 +139,43 @@ class SettingView extends GetView<SettingController> {
 
   Widget _logouButton() {
     return TextButton(
-        onPressed: () {
-          DoUserService.removeUserInfo();
-          // Get.back();
-          Get.offAllNamed("/tabs");
-          EasyLoading.showSuccess("退出登录");
-        },
-        child: Text("退出账号",
-            style: TextStyle(
-                fontSize: DoScreenAdapter.fs(14), color: DoColors.black51)));
-    ;
+      onPressed: () {
+        _showLogoutAlertDialog();
+      },
+      child: Text("退出账号",
+          style: TextStyle(
+              fontSize: DoScreenAdapter.fs(14), color: DoColors.black51)),
+    );
+  }
+
+  void _showLogoutAlertDialog() async {
+    await showCupertinoDialog(
+      context: Get.context!,
+      builder: (BuildContext context) {
+        return CupertinoAlertDialog(
+          title: const Text(
+            "温馨提示",
+          ),
+          content: const Text("确定要退出登录么？"),
+          actions: [
+            TextButton(
+              onPressed: () {
+                DoUserService.removeUserInfo();
+                // Get.back();
+                Get.offAllNamed("/tabs");
+                EasyLoading.showSuccess("退出登录");
+              },
+              child: const Text("确定",
+                  style: TextStyle(fontSize: 14, color: DoColors.theme)),
+            ),
+            TextButton(
+                onPressed: () {
+                  Get.back();
+                },
+                child: const Text("取消"))
+          ],
+        );
+      },
+    );
   }
 }
